// npm-packages/convex/src/cli/lib/generatedLogStreamApi.ts
export interface components {
    schemas: {
        ConsoleLogEvent: {
            function: {
                cached?: boolean | null;
                component_path?: string | null;
                mutation_queue_length?: number | null;
                mutation_retry_count?: number | null;
                path: string;
                request_id: string;
                type: string;
            };
            is_truncated: boolean;
            log_level: string;
            message: string;
            system_code?: string | null;
            /** Format: int64 */
            timestamp: number;
        };
        DeploymentAuditLogEvent: {
            audit_log_action: string;
            audit_log_metadata: string;
            /** Format: int64 */
            timestamp: number;
        };
        FunctionExecutionEvent: {
            error_message?: string | null;
            /** Format: int64 */
            execution_time_ms: number;
            function: {
                cached?: boolean | null;
                component_path?: string | null;
                mutation_queue_length?: number | null;
                mutation_retry_count?: number | null;
                path: string;
                request_id: string;
                type: string;
            };
            occ_info?: null | {
                document_id?: string | null;
                /** Format: int64 */
                retry_count: number;
                table_name?: string | null;
                write_source?: string | null;
            };
            scheduler_info?: null | {
                job_id: string;
            };
            status: string;
            /** Format: int64 */
            timestamp: number;
            usage: {
                /** Format: int64 */
                action_memory_used_mb?: number | null;
                /** Format: int64 */
                database_read_bytes: number;
                /** Format: int64 */
                database_read_documents: number;
                /** Format: int64 */
                database_write_bytes: number;
                /** Format: int64 */
                file_storage_read_bytes: number;
                /** Format: int64 */
                file_storage_write_bytes: number;
                /** Format: int64 */
                vector_storage_read_bytes: number;
                /** Format: int64 */
                vector_storage_write_bytes: number;
            };
        };
        LogStreamEvent: (components["schemas"]["ConsoleLogEvent"] & {
            /** @enum {string} */
            topic: "console";
        }) | (components["schemas"]["VerificationEvent"] & {
            /** @enum {string} */
            topic: "verification";
        }) | (components["schemas"]["FunctionExecutionEvent"] & {
            /** @enum {string} */
            topic: "function_execution";
        }) | (components["schemas"]["DeploymentAuditLogEvent"] & {
            /** @enum {string} */
            topic: "audit_log";
        }) | (components["schemas"]["SchedulerStatsEvent"] & {
            /** @enum {string} */
            topic: "scheduler_stats";
        }) | (components["schemas"]["ScheduledJobLagEvent"] & {
            /** @enum {string} */
            topic: "scheduled_job_lag";
        });
        ScheduledJobLagEvent: {
            /** Format: int64 */
            lag_seconds: number;
            /** Format: int64 */
            timestamp: number;
        };
        SchedulerStatsEvent: {
            /** Format: int64 */
            lag_seconds: number;
            /** Format: int64 */
            num_running_jobs: number;
            /** Format: int64 */
            timestamp: number;
        };
        VerificationEvent: {
            message: string;
            /** Format: int64 */
            timestamp: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type ConsoleLogEvent = components['schemas']['ConsoleLogEvent'];
export type DeploymentAuditLogEvent = components['schemas']['DeploymentAuditLogEvent'];
export type FunctionExecutionEvent = components['schemas']['FunctionExecutionEvent'];
export type LogStreamEvent = components['schemas']['LogStreamEvent'];
export type ScheduledJobLagEvent = components['schemas']['ScheduledJobLagEvent'];
export type SchedulerStatsEvent = components['schemas']['SchedulerStatsEvent'];
export type VerificationEvent = components['schemas']['VerificationEvent'];